feat: change to global components
