+++ frontend/src/components/CreateBoard/SplitBoard/SubTeamsTab/UsersBox/index.tsx
@@ -1,0 +1,46 @@
+import { User } from '@/types/user/user';
+import Text from '@/components/Primitives/Text';
+import BoxContainer from '../BoxContainer';
+import DropdownStakeholders from '../DropdownUsers';
+
+type UsersBoxProps = {
+  haveError: boolean;
+  participants: User[];
+  title: string;
+};
+
+const UsersNames = ({ haveError, participants, title }: UsersBoxProps) => (
+  <BoxContainer color="white">
+    <Text color="primary300" size="xs" css={{ pb: '$2', textAlign: 'start' }}>
+      {title}
+    </Text>
+    <Text
+      css={{
+        textOverflow: 'ellipsis',
+        whiteSpace: 'nowrap',
+        overflow: 'hidden',
+        textAlign: 'start',
+      }}
+      size="md"
+    >
+      {!haveError &&
+        participants &&
+        participants.length > 0 &&
+        participants.map((value) => `${value.firstName} ${value.lastName}`).join(', ')}
+    </Text>
+  </BoxContainer>
+);
+
+const UsersBox = ({ haveError, participants, title }: UsersBoxProps) => (
+  <>
+    {participants.length > 0 ? (
+      <DropdownStakeholders users={participants}>
+        <UsersNames haveError={haveError} participants={participants} title={title} />
+      </DropdownStakeholders>
+    ) : (
+      <UsersNames haveError={haveError} participants={participants} title={title} />
+    )}
+  </>
+);
+
+export default UsersBox;
