+++ frontend/src/components/Users/UsersList/partials/ListOfCards/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useRef, useState } from 'react';
 import Flex from '@/components/Primitives/Flex';
 import { UserWithTeams } from '@/types/user/user';
 import Text from '@/components/Primitives/Text';
-import SearchInput from '@/components/Teams/CreateTeam/ListMembers/SearchInput';
+import SearchInput from '@/components/Teams/CreateTeam/ListMembersDialog/SearchInput';
 import { useSetRecoilState } from 'recoil';
 // import { usersWithTeamsState } from '@/store/user/atoms/user.atom';
 import { useInfiniteQuery } from 'react-query';
